Bug 1999243 - Remove redundant read/execute memory commits. r=bobowen
As identified by :bobowen, when using a shared mapping it is redundant
to commit memory through multiple views. Committing memory in one of the
views propagates commitment to the other views.

This patch removes the redundant commits and adds a test that ensures
that we can rely on this behavior.

Differential Revision: https://phabricator.services.mozilla.com/D271988

UltraBlame original commit: 7c2217073877a560a5bb9466ee0bf68cb669714e