Bug 1730749 - Rework handling of displayport clips for sticky content. r=mstange
To avoid sticky content checkerboarding when we've scrolled farther
than the displayport, we do not apply the displayport clip to it.

This was previously implemented in a hacky way in ClipManager, but
the introduction of sticky ASRs complicates that implementation.

Instead, this patch implements the omission of displayport clips
for sticky content in the Gecko display list itself.

For simplicity, only the case where the displayport clip is the
first element in a sticky item's clip chain is handled.

Differential Revision: https://phabricator.services.mozilla.com/D254156

UltraBlame original commit: 2d9bc48765ec20ad6f4a402bc68f00fae7fad662