Bug 1996608 - [devtools] Display ::view-transition* pseudo elements in markup view. r=devtools-reviewers,ochameau.
This handles all the anonymous content pseudo elements for view transition.
Those elements are a bit different from the pseudo elements we had to handle
until now.
First, they don't have a special nodeName (e.g. for `::before`, we have a `_moz_generated_content_before`
nodeName).
Instead, they have a type attribute, and a name attribute (except `::view-transition`).
So we need to check the type attribute to see if those elements are view transition
pseudo elements.
Those elements also have a common root, `<div type=":-moz-snapshot-containing-block">`
that we don't want to display in the markup view, as its' not exposed in CSS, it's
just an implementation detail.
To handle that, we add a new walker filter result, `FILTER_ACCEPT_CHILDREN`, that
will not be returned, but will have its children evaluated.
This impact the `WalkerActor#rawParentNode` function as we don't want to return
nodes that would have been ignored.

A few test cases are added to make sure the pseudo elements are properly displayed
in the markup view, in the breadcrumb and in the highlighter status bar.

Differential Revision: https://phabricator.services.mozilla.com/D270294

UltraBlame original commit: f6a9102528e3e253d4a6316b4513ed7a2d2e6dd1