Bug 2000330: Use a sans-serif font to avoid fuzzy failure in WPT cache-miss-001.html. r=layout-reviewers,emilio
The green "hi" text on green background in this testcase is not meant to be
visible at all. But the "h" character's top-left-serif draws a little bit of
antialiased fringe barely outside of the green square, which causes a spurious
test failure. Using a sans-serif font will avoid that decoration and prevent
this spurious failure.

Differential Revision: https://phabricator.services.mozilla.com/D272731

UltraBlame original commit: 630dc93241baecfeaf31d741eb1ca8bb4d4b8322