Bug 2000003 - Remove public version nsISearchService.getEnginesByExtensionID. r=search-reviewers,scunnane
This is no longer used outside the service, so we can make it fully private. The existing uses
are within the run-on-idle code which is also guarenteed to be after initialisation, hence we don't
need to await init in those cases.

Differential Revision: https://phabricator.services.mozilla.com/D272619

UltraBlame original commit: e8736e85c8149742ead0075c1a2ef029de29a90e