Bug 1999383 - IPDL serialization of JS values. r=nika,sfink
This patch implements serialization and deserialization of JS values using IPDL.
This is similar to the structured clone algorithm, but the target is chrome JS,
so it does not need to be standards compliant. The idea is that by serializing
to a data type, it is easier to write a type checking algorithm in later patches
in bug 1885221 by traversing the data structure. This also allows us to tightly
control what is allowed to be sent.

This patch removes MMPrinter::Print logging for JS IPC messages, but a more
sophisticated logging will be added in bug 1885221 part 5, once the type
infrastructure is in place.

Differential Revision: https://phabricator.services.mozilla.com/D227393

UltraBlame original commit: ec02c8994cc8f82bdde5221c5a35659015f82a9d