Bug 1997530 - skip test that requires detouring a B instruction on Windows AArch64 r=yjuglaret,win-reviewers
TrackPopupMenu() on ARM uses a B instruction, which is PC-relative. We
don't actually need to detour this, so I think it's OK to just skip this
test on ARM64.

I originally went down the path of making us support B instructions, since
they don't set a register as a side effect. However, this is harder than
I thought as explained in the PatcherDetour.h code - since we use a two-pass
system, on the first pass we assume the remote offset is 0, and that may mean
the offset is out of the 26-bit range in the B instruction.

Since I had already done the work of splitting up the BL and B instructions
in gPCRelTests I went ahead and kept those so it will be (slightly) easier
to implement this in the future if we need to.

While doing this work, I noticed that the already-existing BUncondImmDecode
was incorrect; per the ARM64 documentation the offset to jump to is the
26 bits in the instruction multipled by 4. ( https://developer.arm.com/documentation/ddi0602/2025-03/Base-Instructions/B--Branch- )
We were only using this in Clear4BytePatch ( https://searchfox.org/firefox-main/rev/1a8c62b86277005f907151bc5389cf5c5091e76f/toolkit/xre/dllservices/mozglue/interceptor/PatcherDetour.h#373-374 ), so I'm guessing this didn't come up much.

Differential Revision: https://phabricator.services.mozilla.com/D271077

UltraBlame original commit: 709b0a06f540dec901bf9297cd8e3ce4022a3c1a