Bug 2001613. Don't put ASRs into mActiveScrolledRoots multiple times. r=layout-reviewers,emilio
This doesn't cause problems, but we don't need to do it, and we can clean up these functions a bit to be more clear.

Besides the only functional change of this patch (the first line of the commit message) this patch also renames the functions to GetOrCreate so it is more clear what they do.

Differential Revision: https://phabricator.services.mozilla.com/D273572

UltraBlame original commit: b151f4bb267b1d86db1fceb8b7591a609a0a4ad9