Bug 2002397 - Add CallStackFrameInfo struct for the JitcodeGlobalEntry::callStackAtAddr return values r=iain
Previously, we were returning multiple return arguments for the label
and sourceId return values. It's better to use a struct to return a
single argument with these values instead, so it'll be easier to extend
in the future.

I would like to add line/column information for the stack frames in the
later patches. That's why having this will be beneficial for that patch.

Differential Revision: https://phabricator.services.mozilla.com/D274056

UltraBlame original commit: 4e4bbc433f22e508cbe250346a453b8c1fb83ed6